RootStock-NG: migrate all scripts from the old rootstock script

+++ base_image.sh
@@ -25,6 +25,7 @@ HOST_ARCH=$(uname -m)
 TIME=$(date +%Y-%m-%d)
 
 unset USE_OEM
+unset ROOTSTOCKNG
 
 MINIMAL="-minimal"
 
@@ -75,6 +76,41 @@ function dl_rootstock {
 }
 
 function minimal_armel {
+	if [ "${ROOTSTOCKNG}" ] ; then
+		rm -f "${DIR}/.project" || true
+
+		pkgs="${MINIMAL_APT}${EXTRA}"
+
+		base_pkg_list=$(echo ${pkgs} | sed -e 's/,/ /g')
+
+		tempdir=$(mktemp -d)
+
+		cat > ${DIR}/.project <<-__EOF__
+			tempdir="${tempdir}"
+			distro="${distro}"
+
+			release="${DIST}"
+			dpkg_arch="${ARCH}"
+
+			apt_proxy="${apt_proxy}"
+			base_pkg_list="${base_pkg_list}"
+
+			image_hostname="${FQDN}"
+
+			user_name="${USER_LOGIN}"
+			full_name="${USER_NAME}"
+			password="${USER_PASS}"
+
+			chroot_ENABLE_DEB_SRC="${chroot_ENABLE_DEB_SRC}"
+
+			chroot_KERNEL_HTTP_DIR="${chroot_KERNEL_HTTP_DIR}"
+
+		__EOF__
+
+		cat ${DIR}/.project
+
+		/bin/bash -e "${DIR}/RootStock-NG.sh" || { exit 1 ; }
+	else
 rm -f ${DIR}/deploy/arm*-rootfs-*.tar || true
 rm -f ${DIR}/deploy/vmlinuz-* || true
 rm -f ${DIR}/deploy/initrd.img-* || true
@@ -96,9 +132,11 @@ sudo ${DIR}/git/project-rootstock/rootstock  --imagesize ${IMAGESIZE} --fqdn ${F
 --seed ${MINIMAL_APT}${EXTRA} ${MIRROR} --components "${COMPONENTS}" \
 --dist ${DIST} --script ${DIR}/tools/${FIXUPSCRIPT} \
 --apt-upgrade --arch=${ARCH}
+fi
 }
 
 function compression {
+if [ ! "${ROOTSTOCKNG}" ] ; then
 	rm -rf ${DIR}/deploy/${TIME}/$BUILD || true
 	mkdir -p ${DIR}/deploy/${TIME}/$BUILD
 
@@ -128,10 +166,14 @@ function compression {
 	fi
 
 	cd ${DIR}/deploy/
+else
+	echo "later"
+fi
 }
 
 #12.10
 function quantal_release {
+	distro="ubuntu"
 	reset_vars
 	DIST="quantal"
 
@@ -146,6 +188,7 @@ function quantal_release {
 
 #13.04
 function raring_release {
+	distro="ubuntu"
 	reset_vars
 	DIST="raring"
 
@@ -159,13 +202,14 @@ function raring_release {
 }
 
 function squeeze_release {
+	distro="debian"
 	reset_vars
 	DIST=squeeze
 
 	EXTRA=",isc-dhcp-client,uboot-mkimage,${DEBIAN_ONLY}"
 	USER_LOGIN="debian"
 	FIXUPSCRIPT="fixup-debian-base.sh"
-	MIRROR=$MIRROR_DEB
+	MIRROR="${MIRROR_DEB}"
 	COMPONENTS="${DEB_COMPONENTS}"
 	BUILD="${SQUEEZE_CURRENT}${MINIMAL}-${ARCH}-${TIME}"
 	minimal_armel
@@ -173,13 +217,14 @@ function squeeze_release {
 }
 
 function wheezy_release {
+	distro="debian"
 	reset_vars
 	DIST=wheezy
 
 	EXTRA=",u-boot-tools,${DEBIAN_ONLY}"
 	USER_LOGIN="debian"
 	FIXUPSCRIPT="fixup-debian-base.sh"
-	MIRROR=$MIRROR_DEB
+	MIRROR="${MIRROR_DEB}"
 	COMPONENTS="${DEB_COMPONENTS}"
 	BUILD="${WHEEZY_CURRENT}${MINIMAL}-${ARCH}-${TIME}"
 	minimal_armel
@@ -187,13 +232,14 @@ function wheezy_release {
 }
 
 function sid_release {
+	distro="debian"
 	reset_vars
 	DIST=sid
 
 	EXTRA=",u-boot-tools,${DEBIAN_ONLY}"
 	USER_LOGIN="debian"
 	FIXUPSCRIPT="fixup-debian-base.sh"
-	MIRROR=$MIRROR_DEB
+	MIRROR="${MIRROR_DEB}"
 	COMPONENTS="${DEB_COMPONENTS}"
 	BUILD="${DIST}${MINIMAL}-${ARCH}-${TIME}"
 	minimal_armel
@@ -218,9 +264,15 @@ if [ -f ${DIR}/release ] ; then
 		MIRROR_UBU="--mirror http://ports.ubuntu.com/ubuntu-ports/"
 		MIRROR_DEB="--mirror http://ftp.us.debian.org/debian/"
 	fi
+
+	chroot_ENABLE_DEB_SRC="enable"
 fi
 
-dl_rootstock
+ROOTSTOCKNG=1
+
+if [ ! "${ROOTSTOCKNG}" ] ; then
+	dl_rootstock
+fi
 
 ARCH=armel
 if [ "-${HOST_ARCH}-" == "-armv7l-" ] ; then
